Refactored some handlers and added comments to classes/methods.
