+++ source/GameInterface/Services/Kingdoms/Data/ExpelClanFromKingdomDecisionData.cs
@@ -7,6 +7,9 @@ using TaleWorlds.CampaignSystem.Election;
 
 namespace GameInterface.Services.Kingdoms.Data
 {
+    /// <summary>
+    /// Class for serializing <see cref="ExpelClanFromKingdomDecision"> class.
+    /// </summary>
     [ProtoContract(SkipConstructor = true)]
     public class ExpelClanFromKingdomDecisionData : KingdomDecisionData
     {
@@ -24,6 +27,7 @@ namespace GameInterface.Services.Kingdoms.Data
             OldKingdomId = oldKingdomId;
         }
 
+        /// <inheritdoc/>
         public override bool TryGetKingdomDecision(IObjectManager objectManager, out KingdomDecision kingdomDecision)
         {
             if (!TryGetProposerClanAndKingdom(objectManager, out Clan proposerClan, out Kingdom kingdom) ||
