+++ source/GameInterface/Services/Kingdoms/Data/KingSelectionKingdomDecisionData.cs
@@ -9,6 +9,9 @@ using TaleWorlds.CampaignSystem.Election;
 
 namespace GameInterface.Services.Kingdoms.Data
 {
+    /// <summary>
+    /// Class for serializing <see cref="KingSelectionKingdomDecision"> class.
+    /// </summary>
     [ProtoContract(SkipConstructor = true)]
     public class KingSelectionKingdomDecisionData : KingdomDecisionData
     {
@@ -21,6 +24,7 @@ namespace GameInterface.Services.Kingdoms.Data
             ClanToExcludeId = clanToExcludeId;
         }
 
+        /// <inheritdoc/>
         public override bool TryGetKingdomDecision(IObjectManager objectManager, out KingdomDecision kingdomDecision)
         {
             if (!TryGetProposerClanAndKingdom(objectManager, out Clan proposerClan, out Kingdom kingdom) ||
