+++ source/GameInterface/Services/Kingdoms/Data/SettlementClaimantPreliminaryDecisionData.cs
@@ -10,6 +10,9 @@ using TaleWorlds.CampaignSystem.Settlements;
 
 namespace GameInterface.Services.Kingdoms.Data
 {
+    /// <summary>
+    /// Class for serializing <see cref="SettlementClaimantPreliminaryDecision"> class.
+    /// </summary>
     [ProtoContract(SkipConstructor = true)]
     public class SettlementClaimantPreliminaryDecisionData : KingdomDecisionData
     {
@@ -26,6 +29,7 @@ namespace GameInterface.Services.Kingdoms.Data
             OwnerClanId = ownerClanId;
         }
 
+        /// <inheritdoc/>
         public override bool TryGetKingdomDecision(IObjectManager objectManager, out KingdomDecision kingdomDecision)
         {
             if (!TryGetProposerClanAndKingdom(objectManager, out Clan proposerClan, out Kingdom kingdom) ||
