+++ source/GameInterface/Services/Kingdoms/Messages/LocalDecisionRemoved.cs
@@ -2,6 +2,9 @@
 
 namespace GameInterface.Services.Kingdoms.Messages
 {
+    /// <summary>
+    /// Event that is handled on server side, when Kingdom.RemoveDecision method is called.
+    /// </summary>
     public record LocalDecisionRemoved : IEvent
     {
         public string KingdomId { get; }
