+++ source/GameInterface/Services/Kingdoms/Messages/RemoveDecision.cs
@@ -2,6 +2,9 @@
 
 namespace GameInterface.Services.Kingdoms.Messages
 {
+    /// <summary>
+    /// Event that is handled on client side, when Server sends NetworkRemoveDecision message to clients.
+    /// </summary>
     public class RemoveDecision: IEvent
     {
         public string KingdomId { get; }
